synchronizer/requirements.txt
kafka-python==2.0.2
Pillow==8.1.2